pyproject.toml
[tool.poetry]
name = "mytempura"
version = "0.1.0"
description = ""
authors = ["Hoang Do <tuanhoang2do2311@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.4"
coverage = "^6.5.0"
firebase-admin = "^6.1.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.25.1"
celery = "^5.2.7"
redis = "^4.5.4"
pyjwt = "^2.7.0"
django-cors-headers = "^4.0.0"
watchdog = "^3.0.0"
django-jazzmin = "^2.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
